+++ src/indicators/simple_moving_average.rs
@@ -41,15 +41,15 @@ use serde::{Deserialize, Serialize};
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Debug, Clone)]
 pub struct SimpleMovingAverage {
-    length: u32,
+    length: usize,
     index: usize,
-    count: u32,
+    count: usize,
     sum: f64,
     vec: Vec<f64>,
 }
 
 impl SimpleMovingAverage {
-    pub fn new(length: u32) -> Result<Self> {
+    pub fn new(length: usize) -> Result<Self> {
         match length {
             0 => Err(Error::from_kind(ErrorKind::InvalidParameter)),
             _ => {
@@ -58,14 +58,14 @@ impl SimpleMovingAverage {
                     index: 0,
                     count: 0,
                     sum: 0.0,
-                    vec: vec![0.0; length as usize],
+                    vec: vec![0.0; length],
                 };
                 Ok(indicator)
             }
         }
     }
 
-    pub fn length(&self) -> u32 {
+    pub fn length(&self) -> usize {
         self.length
     }
 }
@@ -77,7 +77,7 @@ impl Next<f64> for SimpleMovingAverage {
         let old_val = self.vec[self.index];
         self.vec[self.index] = input;
 
-        self.index = if self.index + 1 < self.length as usize {
+        self.index = if self.index + 1 < self.length {
             self.index + 1
         } else {
             0
@@ -105,7 +105,7 @@ impl Reset for SimpleMovingAverage {
         self.index = 0;
         self.count = 0;
         self.sum = 0.0;
-        for i in 0..(self.length as usize) {
+        for i in 0..self.length {
             self.vec[i] = 0.0;
         }
     }
